refactor/54: GeoService 캐시 로직 분리
